split trotterization into its own file
