+++ src/qpe_toolbox/hamiltonian/__init__.py
@@ -10,3 +10,4 @@
 
 from .chemistry import chemistry_hamiltonian, do_pyscf, make_qubit_hamiltonian
 from .hamiltonian import Hamiltonian, do_dmrg, heisenberg_hamiltonian
+from .trotterization import trotter_approx_as_MPO
