+++ tests/test_mpomps_tools.py
@@ -3,7 +3,8 @@
 import numpy as np
 import quimb.tensor as qtn
 
-from qpe_toolbox.tensor import kron_mpos, kron_mps
+from qpe_toolbox.hamiltonian import heisenberg_hamiltonian
+from qpe_toolbox.tensor import kron_mpos, kron_mps, state_preparation_mpo
 
 # define
 
@@ -40,7 +41,31 @@ def test_kronmpos():
     assert abs((myId3_bis - Id3).norm()) < 1e-10
 
 
+def test_state_preparation_mpo():
+    # state_preparation_mpo expects the boundary-tensor leg order DMRG2
+    # produces (phys, bond) / (bond, phys) -- not MPS_rand_state's (bond, phys)
+    ham = heisenberg_hamiltonian(3)
+    dmrg = qtn.DMRG2(
+        ham.to_mpo(), p0=qtn.MPS_rand_state(ham.n_qubits, bond_dim=2, seed=42)
+    )
+    dmrg.solve(max_sweeps=8, bond_dims=16, verbosity=0, cutoffs=1e-10)
+    gs = dmrg.state
+    gs_vec = gs.to_dense().reshape(-1)
+
+    dense = state_preparation_mpo(state_mps=gs).to_dense()
+
+    n = ham.n_qubits
+    e0 = np.zeros(2**n, dtype=complex)
+    e0[0] = 1.0
+    assert np.allclose(dense @ e0, 2**n * gs_vec, atol=1e-8)
+
+    e1 = np.zeros(2**n, dtype=complex)
+    e1[1] = 1.0
+    assert np.allclose(dense @ e1, 0, atol=1e-8)
+
+
 # run
 if __name__ == "__main__":
     test_kronmps()
     test_kronmpos()
+    test_state_preparation_mpo()
